+++ plumbing/transport/http/git_upload_pack_test.go
@@ -7,7 +7,7 @@ import (
 
 	. "gopkg.in/check.v1"
 	"gopkg.in/src-d/go-git.v4/plumbing"
-	"gopkg.in/src-d/go-git.v4/plumbing/client/common"
+	"gopkg.in/src-d/go-git.v4/plumbing/transport/common"
 )
 
 type RemoteSuite struct {
