+++ plumbing/transport/ssh/auth_method_test.go
